#41007 Move cache updater into separate file, rename AdminSearchService to SearchRepository
